requirements.txt
PyJWT>=1.1.0
requests>=2.0.0
cryptography>=3.0.0